extract Borders
