+++ samples/showcase/src/main/kotlin/org/jetbrains/jewel/samples/showcase/components/Borders.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.jewel.samples.standalone.view.component
+package org.jetbrains.jewel.samples.showcase.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -28,7 +28,7 @@ import org.jetbrains.jewel.ui.component.Text
 import org.jetbrains.jewel.ui.theme.colorPalette
 
 @Composable
-internal fun Borders() {
+public fun Borders() {
     GroupHeader("Group header")
     Text("This is a group header example")
